import debugproxy
This is a source dump from the debugproxy repository minus the
deployment ansible code and load testing code.

The history prior to this has been removed as it was too difficult to
check if there was any sensitive information in it. Maybe it can be
added later.

Most of this code was writen 2 to 3 years ago and there have been a
handful of commits since then fixing minor bugs and some minor updates
in the deployment scripts.